+++ pulsar/utils/importer.py
@@ -88,3 +88,5 @@ def import_system_file(mod, add_to_path=True):
         else:
             mod_name = '.'.join(names)
         return import_module(mod_name)
+    else:
+        return import_module(mod)